Consolidate BaseTestAESCipherInputStreamExceptions and BaseTestAESGCMCICOWithGCMAndAAD tests
This change consolidates the tests previously associated with BaseTestAESCipherInputStreamExceptions into the parameterized TestAESCipherInputStreamExceptions class, and the tests from BaseTestAESGCMCICOWithGCMAndAAD into the parameterized TestAESGCMCICOWithGCMAndAAD class.

Fixes: https://github.ibm.com/runtimes/jit-crypto/issues/1083
