--- src/test/java/ibm/jceplus/junit/openjceplus/multithread/TestAESGCMCICOWithGCMAndAAD.java
@@ -1,25 +1,0 @@
-/*
- * Copyright IBM Corp. 2023, 2024
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms provided by IBM in the LICENSE file that accompanied
- * this code, including the "Classpath" Exception described therein.
- */
-
-package ibm.jceplus.junit.openjceplus.multithread;
-
-import ibm.jceplus.junit.base.BaseTestAESGCMCICOWithGCMAndAAD;
-import ibm.jceplus.junit.openjceplus.Utils;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.TestInstance.Lifecycle;
-
-@TestInstance(Lifecycle.PER_CLASS)
-public class TestAESGCMCICOWithGCMAndAAD extends BaseTestAESGCMCICOWithGCMAndAAD {
-
-    @BeforeAll
-    public void beforeAll() {
-        Utils.loadProviderTestSuite();
-        setProviderName(Utils.TEST_SUITE_PROVIDER_NAME);
-    }
-}
