+++ src/test/java/ibm/jceplus/junit/openjceplusfips/TestAll.java
@@ -17,12 +17,10 @@ import org.junit.platform.suite.api.Suite;
     TestAESCCM2.class,
     TestAESCCMInteropBC.class,
     TestAESCCMParameters.class,
-    TestAESCipherInputStreamExceptions.class,
     TestAESCopySafe.class,
     TestAESGCM_ExtIV.class,
     TestAESGCM_IntIV.class,
     TestAESGCMBufferIV.class,
-    TestAESGCMCICOWithGCMAndAAD.class,
     TestAESGCMLong.class,
     TestAESGCMNonExpanding.class,
     TestAESGCMSameBuffer.class,
