+++ src/test/java/ibm/jceplus/junit/tests/TestAESCipherInputStreamExceptions.java
@@ -1,12 +1,12 @@
 /*
- * Copyright IBM Corp. 2023
+ * Copyright IBM Corp. 2023, 2026
  *
  * This code is free software; you can redistribute it and/or modify it
  * under the terms provided by IBM in the LICENSE file that accompanied
  * this code, including the "Classpath" Exception described therein.
  */
 
-package ibm.jceplus.junit.base;
+package ibm.jceplus.junit.tests;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -16,8 +16,13 @@ import javax.crypto.spec.GCMParameterSpec;
 import javax.crypto.spec.IvParameterSpec;
 import javax.crypto.spec.SecretKeySpec;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInfo;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.params.Parameter;
+import org.junit.jupiter.params.ParameterizedClass;
+import org.junit.jupiter.params.provider.MethodSource;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -29,7 +34,17 @@ import static org.junit.jupiter.api.Assertions.fail;
  * - Make sure other algorithms do not throw exceptions when the stream
  *   calls close() and only throw when a invalid read() errors.
  */
-public class BaseTestAESCipherInputStreamExceptions extends BaseTestJunit5 {
+@Tag(Tags.OPENJCEPLUS_NAME)
+@Tag(Tags.OPENJCEPLUS_FIPS_NAME)
+@Tag(Tags.OPENJCEPLUS_MULTITHREAD_NAME)
+@Tag(Tags.OPENJCEPLUS_FIPS_MULTITHREAD_NAME)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@ParameterizedClass
+@MethodSource("ibm.jceplus.junit.tests.TestArguments#getEnabledProviders")
+public class TestAESCipherInputStreamExceptions extends BaseTest {
+
+    @Parameter(0)
+    TestProvider provider;
 
     static final SecretKeySpec key = new SecretKeySpec(new byte[16], "AES");
     static final GCMParameterSpec gcmspec = new GCMParameterSpec(128, new byte[16]);
@@ -37,8 +52,9 @@ public class BaseTestAESCipherInputStreamExceptions extends BaseTestJunit5 {
     private TestInfo testInfo;
 
     @BeforeEach
-    void init(TestInfo testInfo) {
+    void init(TestInfo testInfo) throws Exception {
         this.testInfo = testInfo;
+        setAndInsertProvider(provider);
     }
 
     /**
